#1 GUI파일호환설정
