+++ .idea/workspace.xml
@@ -1,21 +1,46 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
   <component name="ChangeListManager">
-    <list default="true" id="b0dc4ce5-5407-48b6-8771-575491c48c16" name="Changes" comment="" />
+    <list default="true" id="b0dc4ce5-5407-48b6-8771-575491c48c16" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/braillePrinter.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/braillePrinter.iml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/CreateUserClass.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/CreateUserClass.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/HomeClass.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/HomeClass.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/HomeClass_temp.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/LoginClass.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/LoginClass.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/UserClass.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/UserClass.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/UserFormerInfoClass.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/UserFormerInfoClass.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gui/serverFunction.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/gui/serverFunction.py" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
   </component>
   <component name="ProjectId" id="2JH4w1PvxpSDOwMvZ63psMcZsVP" />
   <component name="ProjectViewState">
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">
-    <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
-    <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
-  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "last_opened_file_path": "C:/Users/JDoubleU/Desktop/Capstone-braille-Printer",
+    "settings.editor.selected.configurable": "com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable"
+  }
+}]]></component>
   <component name="RunManager">
     <configuration name="main" type="PythonConfigurationType" factoryName="Python" nameIsGenerated="true">
       <module name="braillePrinter" />
@@ -49,16 +74,5 @@
       <updated>1671721910218</updated>
     </task>
     <servers />
-  </component>
-  <component name="XDebuggerManager">
-    <breakpoint-manager>
-      <breakpoints>
-        <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
-          <url>file://$PROJECT_DIR$/main.py</url>
-          <line>8</line>
-          <option name="timeStamp" value="1" />
-        </line-breakpoint>
-      </breakpoints>
-    </breakpoint-manager>
   </component>
 </project>
