+++ main.py
@@ -1,24 +1,20 @@
-from LoginClass import Login
-from HomeClass_temp import Home
-from CreateUserClass import CreateUser
-from UserClass import User
-
-from StateClass import State
-import StateClass
-
-import serverFunction
+from src.gui import serverFunction, StateClass
+from src.gui.CreateUserClass import CreateUser
+from src.gui.HomeClass import Home
+from src.gui.LoginClass import Login
+from src.gui.UserClass import User
 
 con = serverFunction.getConnect()
 
 while True:
-  # 로그인  UI 상태 //  로그인 성공, 회원가입, 둘다 아님
+  # 로그인  UI 상태 //  로그인 성공,eInstance.password, con)
+  print("로구인 성공")
   login = Login(con)
   # 성공인 상황 / 회원가입 해야하는 상황 / 아무것도 아닌 상황
   stateInstance = login.start()
-  
+
   if stateInstance.state == StateClass.State.LOGINOK: # 로그인 성공
-    user = User(stateInstance.id, stateInstance.password, con)
-    print("로구인 성공")
+    user = User(stateInstance.id, stateInstance.password)
     home = Home(user)
     
     # homeUI에서 로그아웃을 누르면 다시 if문을 빠져나가서 wile 무한루프로 들어감
