+++ src/gui/HomeClass.py
@@ -1,20 +1,47 @@
 from tkinter import *
 import tkinter.ttk as ttk
 from tkinter import filedialog # 파일 선택 (__all__에 없어서), 파일창
+import tkinter.messagebox as msgbox
 
-import homeFunction
-from UserClass import User
+from src.gui import serverFunction, homeFunction
+from src.gui.UserFormerInfoClass import UserFormerInfo
+
 
 class Home:
   def __init__(self, user):
     self.__user = user
     self.__window=self.__createUI()
-    self.word_list = None
+    
+    self.word_list = []
+    
+    #self.text_place내에 있던 문자들 중에 점자 번역 매칭 안되는 문자들에 대한 text내 인덱스를 가지고 있는 리스트
+    self.incorrect_word_list = []
     
   def __createUI(self):
     root = Tk()
     root.title("braille printer GUI")
     root.resizable(False, False) # x, y값 변경 불가(창 크기 변경 불가)
+
+   # Menu
+   # menu = Menu(root, background='blue')
+    menu = Menu(root)
+    menu.config(background='blue')
+    
+    menu_1 = Menu(menu, tearoff=0)
+    menu_1.add_command(label="인쇄정보", command=self.print_information_function)
+    menu_1.add_separator()
+    menu_1.add_command(label="내 정보 더보기", command=self.more_my_information_function)
+    menu.add_cascade(label="내 정보", menu=menu_1)
+
+    menu_2 = Menu(menu, tearoff=0)
+    menu_2.add_command(label="로그아웃", command=self.exit)
+    menu_2.add_command(label="회원 탈퇴", command=self.user_delete)
+    menu_2.add_separator()
+    menu_2.add_command(label="종료", command= self.quit)
+    menu.add_cascade(label="더보기", menu=menu_2)
+
+    root.config(menu=menu)
+
 
     #Label Frame
     labelFrame = Frame(root)
@@ -36,30 +63,86 @@ class Home:
     scrollbar.pack(side='right', fill='y')
 
     self.text_place = Text(text_frame, height=20, yscrollcommand=scrollbar.set, state="normal")
+    self.text_place.mark_set("","0.0")
     self.text_place.pack(side='left', fill='both', expand='True')
     scrollbar.config(command=self.text_place.yview)
 
-    # Mode Frame {작성 모드, 파일 모드}
+  # Text Button Frame {Text_Button_Frame={left_button_frame, rigt_button_frame} }
+    Text_Button_Frame = Frame(root)
+    Text_Button_Frame.pack(fill='both', padx=5, pady=5)
+    
+    # -------------- left_button_frame--------------
+    left_button_frame = Frame(Text_Button_Frame)
+    left_button_frame.pack(fill='x', padx=5, pady=5, side='left')
+    
+    # -------------- left_button_frame 버튼 1행 --------------
+    double_quotationLeft_button = Button(left_button_frame, width=2, height=1, text="“\n여는큰따옴표", command=lambda :self.text_place.insert(END, "“"))
+    double_quotationRight_button = Button(left_button_frame, width=2, height=1, text="”", command=lambda :self.text_place.insert(END, "”"))
+    single_quotationLeft_button = Button(left_button_frame, width=2, height=1, text="‘", command=lambda :self.text_place.insert(END, "‘"))
+    single_quotationRight_button = Button(left_button_frame, width=2, height=1, text="’", command=lambda :self.text_place.insert(END, "’"))
+    comma_button = Button(left_button_frame, width=2, height=1, text=",", command=lambda :self.text_place.insert(END, ","))
+    
+   # -------------- left_button_frame 버튼 2행 --------------
+
+    hyphen_button = Button(left_button_frame, width=2, height=1,text="‐", command=lambda :self.text_place.insert(END, "‐"))
+    dash_button = Button(left_button_frame, width=2, height=1,text="–", command=lambda :self.text_place.insert(END, "–"))
+    long_dash_button = Button(left_button_frame, width=2, height=1,text="―", command=lambda :self.text_place.insert(END, "―"))
+    minus_button = Button(left_button_frame, width=2, height=1,text="−", command=lambda :self.text_place.insert(END, "−"))
+    
+    # -------------- left_button_frame 버튼 배치 --------------
+    double_quotationLeft_button.grid(row=0, column=0, padx=2, pady=2)
+    double_quotationRight_button.grid(row=0, column=1, padx=2, pady=2)
+    single_quotationLeft_button.grid(row=0, column=2, padx=2, pady=2)
+    single_quotationRight_button.grid(row=0, column=3, padx=2, pady=2)
+    comma_button.grid(row=0, column=4, padx=2, pady=2)
+    hyphen_button.grid(row=1, column=0, padx=2, pady=2)
+    dash_button.grid(row=1, column=1, padx=2, pady=2)
+    long_dash_button.grid(row=1, column=2, padx=2, pady=2)
+    minus_button.grid(row=1, column=3, padx=2, pady=2)
+
+    # -------------- right_button_frame--------------
+    rigt_button_frame = Frame(Text_Button_Frame)
+    rigt_button_frame.pack(fill='x', padx=5, pady=5, side='right')
+    
+    # -------------- rigt_button_frame 버튼 1행 --------------
+    check_button = Button(rigt_button_frame, width=5, height=1, text="검사", command=self.check_braille_grammar)
+    left_arrow_button = Button(rigt_button_frame, width=2, height=1, text="<", command=self.go2Left)
+    right_arrow_button = Button(rigt_button_frame, width=2, height=1, text=">", command=self.go2Right)
+    Braille_button = Button(rigt_button_frame, width=5, height=1, text="점역", command=self.braille_review)
+
+    # -------------- rigt_button_frame 버튼 배치 --------------
+    check_button.grid(row=0, column=0, padx=2 ,pady=2)
+    left_arrow_button.grid(row=0, column=1, padx=2 ,pady=2)
+    right_arrow_button.grid(row=0, column=2, padx=2 ,pady=2)
+    Braille_button.grid(row=0, column=3, padx=2 ,pady=2)
+
+  # -------------- braille Fraem --------------
+
+    braille_frame = LabelFrame(root, text="점자 번역")
+    braille_frame.pack(fill='both', padx=5, pady=5)
+
+    scrollbar = Scrollbar(braille_frame)
+    scrollbar.pack(side='right', fill='y')
+
+    self.braille_place = Text(braille_frame, height=20, yscrollcommand=scrollbar.set, state="disabled")
+    self.braille_place.pack(side='left', fill='both', expand='True')
+    scrollbar.config(command=self.braille_place.yview)
+
+  # -------------- Mode Frame {작성 모드, 파일 모드} --------------
 
     mode_frame = LabelFrame(root, text="모드 선택")
     mode_frame.pack(fill='x', padx=5, pady=5)
-
-    #######################################################
- 
-        
-    #######################################################
     
     self.write_var = IntVar()
     write_radio_button = Radiobutton(mode_frame, text="작성모드", value=1, variable= self.write_var, command=self.mode_function)
     file_radio_button = Radiobutton(mode_frame, text="파일모드", value=2, variable= self.write_var, command=self.mode_function)
 
 
-
     write_radio_button.select()
     file_radio_button.pack(side='right')
     write_radio_button.pack(side='right')
 
-    # File Select Frame
+  # -------------- File Select Frame --------------
 
     button_Frame = Frame(root)
     button_Frame.pack(fill='x', padx=5, pady=5)
@@ -70,7 +153,7 @@ class Home:
     self.file_unselection_button.pack(side="right", padx=3)
     self.file_selection_button.pack(side='right', padx=3)
 
-    # 진행 상황 Progress Bar Frame
+  # -------------- 진행 상황 Progress Bar Frame --------------
     frame_progress = LabelFrame(root, text='진행상황')
     frame_progress.pack(fill='x', padx=5, pady=5)
 
@@ -78,7 +161,7 @@ class Home:
     progress_bar = ttk.Progressbar(frame_progress, maximum=100, variable=p_var)
     progress_bar.pack(fill='x', padx=5, pady=5)
 
-    # Print  Frame
+  # -------------- Print  Frame --------------
     prit_Frmae = Frame(root)
     prit_Frmae.pack(fill='x', padx=5, pady=5)
     print_button = Button(prit_Frmae,  width = 13,height = 1, text="프린트 시작", command=self.startPrint)
@@ -89,22 +172,56 @@ class Home:
     return root
   
   
+  
+  ############################# menu function ###################################
+  
+  # 이전 프린트 목록 보여주기
+  def print_information_function(self):
+    # 리스트형식이고, 안에는 튜플들(이전에 인쇄정보들에 대한)이 있응
+    #historyList = serverFunction.get_user_history()
+    historyList = [{"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}, {"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}, {"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}, {"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}, {"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}, {"title": "test document title", "page": 3, "state": "인쇄중", "submit_at": "2023-02-23T15:42:55.000Z"}]
+    
+    # 못받아오면 경고
+    if historyList is None:
+      msgbox.showerror(title="불러오기 실패", message="불러오기를 실패했습니다.")
+    else:
+      uInfo = UserFormerInfo(historyList)
+      uInfo.start()
+      
+  def more_my_information_function(self):
+    pass
+    
+  def user_delete(self):
+    id = self.__user.getId()
+    response = msgbox.askyesno(title="회원 탈퇴", message=f"{id} 회원을 탈퇴합니다")
+    
+    if(response):
+      isDelUser = serverFunction.user_delete(self.__user)
+      if(isDelUser):
+        msgbox.showinfo(title="탈퇴 성공", message="탈퇴에 성공했습니다")
+        self.exit()
+      else:
+        msgbox.showerror(title="탈퇴 실패", message="탈퇴에 실패했습니다")
+  ############################# mode function ###################################
+
+  
   def mode_function(self):
     num = self.write_var.get()
     
     if num==1: #작성모드
-      self.text_place.config(state="normal")
+      self.text_place.delete("0.0", END)
       self.file_selection_button.config(state="disabled")
       self.file_unselection_button.config(state="disabled")
-      self.text_place.delete("0.0", END)
       
       
     else: # 피일모드
-      self.text_place.delete("0.0", END)
+      self.text_place.delete("0.0", END)	
       self.text_place.config(state="disabled")
       self.file_selection_button.config(state="normal")
       self.file_unselection_button.config(state="normal")
       
+  ############################# getDate function ###################################
+
   def getData(self):
     # 파일의 절대 위치 경로를 리턴받음
     filepath = filedialog.askopenfilename(title="워드 파일 선택", filetypes=(("docx 파일",'*.docx'), ("hwp 파일",'*.hwp'), ("pdf 파일",'*.pdf'), ("txt 파일",'*.txt'), ("모든 파일", "*.*")), initialdir=r"C:\Users\JDoubleU\Desktop")
@@ -124,23 +241,52 @@ class Home:
       print("Unknown extension")
       return;
     
-    self.text_place.config(state='normal')
     # 텍스트 있을 수도 있으니까 삭제하고 집어넣기
     self.text_place.delete("0.0", END)	
+    self.text_place.config(state='normal')
+    
+    #self.wort_list에는 파일에서 or 작성모드에서 작성한 글자들, 대응되는 유니코드 값들이 있음
+    # ex) "abc" -> word_list = [(a, 65),(c, 67),(b, 66)]
     for c in self.word_list:
       self.text_place.insert(END, c[0])
     
-    self.text_place.config(state='disabled')
+  ############################# right button function ###################################
+  def check_braille_grammar(self):
     
-   
-  # 아두이노 구현
-  def startPrint(self):
+    string = self.text_place.get("1.0", END)
+    #워드리스트 최신화(파일들 로딩하고 바꿨을 가능성, 작성모드했으면 아예 워드 리스트에 아무것도 없어서)
+    self.word_list =  homeFunction.get_string(string)
+    print(self.word_list)
+    pass
+    
+  def go2Left(self):
+    pass  
+  
+  def go2Right(self):
     pass
   
+  # 점자 미리보기 구현 함수
+  def braille_review(self):
+    pass
+  
+  ############################# arduino function ###################################
+  def startPrint(self):
+    pass
+
+
+  ############################# process function ###################################  
   def start(self):
     self.__window.mainloop() 
     
+  
+  def exit(self):
+    self.__window.destroy()
+
+  def quit(self):
+    self.__window.quit()
     
-user = User("JW","asdasd",True)
-home = Home(user)
-home.start()
+#####################################################################################
+ 
+# user = User("JW","asdasd",True)
+# home = Home(user)
+# home.start()
